requirements.txt
google_auth_oauthlib
asn1crypto==0.24.0
backports-abc==0.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
cryptography==3.3.2
enum34==1.1.6
future==0.17.1
futures==3.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.19.0
ipaddress==1.0.22
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
python-telegram-bot==12.0.0
rsa==4.1
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
uritemplate==3.0.0
python-telegram-bot
pyrebase
loguru

## uploadFirebase
firebase_admin
jwt